third_party/WebKit/LayoutTests/inspector-protocol/accessibility/accessibility-getNodeWithNoAXNode-expected.txt


result: {
  "accessibilityNode": {
    "nodeId": "0",
    "ignored": true,
    "ignoredReasons": [
      {
        "name": "notRendered",
        "value": {
          "type": "boolean",
          "value": true
        }
      }
    ]
  }
}
